Fixed compile errors for x64
Binaries submitted in 962170
Build time : 9 minutes 28.172000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 962158]

+++ Tr2PickingHelperBatch.h
@@ -111,7 +111,7 @@ public:
 		renderContext.m_esm.ApplyVertexDeclaration( g_debugResourceHelper.GetVertexPosColorDecl() );
 
 		renderContext.SetTopology( Tr2RenderContextEnum::TOP_TRIANGLES );
-		renderContext.DrawPrimitiveUP( m_vertices.size() / 3, &m_vertices.front(), sizeof( Vector4 ) );
+		renderContext.DrawPrimitiveUP( uint32_t( m_vertices.size() / 3 ), &m_vertices.front(), uint32_t( sizeof( Vector4 ) ) );
 	}
 private:
